make examples readable by customers
examples used in testing go to examples/internal
